requirements.txt
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
idna==2.8
notify-run==0.0.13
pushbullet.py==0.11.0
PyQRCode==1.2.1
python-magic==0.4.15
requests==2.22.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.9
websocket-client==0.57.0
